refactor: extract logic into hooks
